Time and Photolysis Fix
Previously, it was easy to confuse JDAY as GMT. It was entered as fractional day GMT, but then internally it was converted to local time. As a result, it was unintentionaly used as GMT for photolysis.

This fix creates GMT and LOCAL time copies of JDAY that can be directly used by ZENITH for example.

I also added debug options to set_up_photol. This makes it easy to see which indices to use for TUV_J.