Fixes and Debug
The photolysis module now provides a check to see if TUV_NEW is
correctly matched to MCM J order. This is important and was done
because values were not matching.

The order of TUVJ to MCMJ mappings were updated.

Also replaced ZENITH function (borrowed from CMAQ) with a ZENITH
function built on TUV sunae.