Improving platform customizabiltiy
Many systems do not have libfl.a installed by default. I have added traditional options
to Makefile.def.in and kpp/src/Makefile so that LDFLAGS at configure time will allow for libfl
to be in a custom locations.

TUV and FASTJX are available as static libraries, which require Position Independent Code (PIC)
options on linux. These options are not required on Mac or other BSD systems that use PIC by
default. I have added checks in configure.ac that propagate to configure to enable PIC
on non-Darwin systems.